+++ deployment/build-s3-dist.sh
@@ -4,11 +4,11 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 # PURPOSE:
-#   Build cloud formation templates for the Media Insights Engine
+#   Build cloud formation templates for the Media Insights on AWS
 #
 # USAGE:
 #  ./build-s3-dist.sh [-h] [-v] [--no-layer] --template-bucket {TEMPLATE_BUCKET} --code-bucket {CODE_BUCKET} --version {VERSION} --region {REGION} --profile {PROFILE}
-#    TEMPLATE_BUCKET should be the name for the S3 bucket location where MIE
+#    TEMPLATE_BUCKET should be the name for the S3 bucket location where MI
 #      cloud formation templates should be saved.
 #    CODE_BUCKET should be the name for the S3 bucket location where cloud
 #      formation templates should find Lambda source code packages.
@@ -224,7 +224,7 @@ echo "mkdir -p $regional_dist_dir"
 mkdir -p "$regional_dist_dir"
 
 echo "------------------------------------------------------------------------------"
-echo "Building MIEHelper package"
+echo "Building MIHelper package"
 echo "------------------------------------------------------------------------------"
 
 cd "$source_dir"/lib/MediaInsightsEngineLambdaHelper || exit 1
@@ -261,15 +261,15 @@ else
   rm -f Media_Insights_Engine*.whl
   cp -R "$source_dir"/lib/MediaInsightsEngineLambdaHelper .
   cd MediaInsightsEngineLambdaHelper/ || exit 1
-  echo "Building MIE Lambda Helper python library"
+  echo "Building MI Lambda Helper python library"
   python3 setup.py bdist_wheel > /dev/null
   cp dist/*.whl ../
   cp dist/*.whl "$source_dir"/lib/MediaInsightsEngineLambdaHelper/dist/
-  echo "MIE Lambda Helper python library is at $source_dir/lib/MediaInsightsEngineLambdaHelper/dist/"
+  echo "MI Lambda Helper python library is at $source_dir/lib/MediaInsightsEngineLambdaHelper/dist/"
   cd "$source_dir"/lib/MediaInsightsEngineLambdaHelper/dist/ || exit 1
   ls -1 "$(pwd)"/*.whl
   if [ $? -ne 0 ]; then
-    echo "ERROR: Failed to build MIE Lambda Helper python library"
+    echo "ERROR: Failed to build MI Lambda Helper python library"
     exit 1
   fi
   cd "$build_dir"/lambda_layer_factory/ || exit 1
@@ -883,8 +883,8 @@ if [ "$global_bucket" != "solutions-reference" ] && [ "$global_bucket" != "solut
   echo "---"
 
   set -x
-  aws s3 sync $global_dist_dir s3://$global_bucket/aws-media-insights-engine/$version/ $(if [ ! -z $profile ]; then echo "--profile $profile"; fi)
-  aws s3 sync $regional_dist_dir s3://${regional_bucket}-${region}/aws-media-insights-engine/$version/ $(if [ ! -z $profile ]; then echo "--profile $profile"; fi)
+  aws s3 sync $global_dist_dir s3://$global_bucket/media-insights-on-aws/$version/ $(if [ ! -z $profile ]; then echo "--profile $profile"; fi)
+  aws s3 sync $regional_dist_dir s3://${regional_bucket}-${region}/media-insights-on-aws/$version/ $(if [ ! -z $profile ]; then echo "--profile $profile"; fi)
   set +x
 
   echo "------------------------------------------------------------------------------"
@@ -893,11 +893,11 @@ if [ "$global_bucket" != "solutions-reference" ] && [ "$global_bucket" != "solut
 
   echo ""
   echo "Template to deploy:"
-  echo "TEMPLATE='"https://"$global_bucket"."$s3domain"/aws-media-insights-engine/"$version"/media-insights-stack.template"'"
+  echo "TEMPLATE='"https://"$global_bucket"."$s3domain"/media-insights-on-aws/"$version"/media-insights-stack.template"'"
 
   # Save the template URI for test automation scripts:
   touch templateUrl.txt
-  echo "https://"$global_bucket"."$s3domain"/aws-media-insights-engine/"$version"/media-insights-stack.template" > templateUrl.txt
+  echo "https://"$global_bucket"."$s3domain"/media-insights-on-aws/"$version"/media-insights-stack.template" > templateUrl.txt
 fi
 
 cleanup