+++ docs/sigv4_get_sample.py
@@ -1,8 +1,8 @@
 ###############################################################################
 #
-# PURPOSE: This code sample shows how to list MIE workflows using a
+# PURPOSE: This code sample shows how to list MI workflows using a
 #   sigv4 signed request to the workflow execution REST API in the Media
-#   Insights Engine.
+#   Insights on AWS.
 #
 ###############################################################################
 #