+++ test/integ/workflowapi/conftest.py
@@ -24,8 +24,8 @@ def testing_env_variables():
             'SAMPLE_TERMINOLOGY': os.environ['SAMPLE_TERMINOLOGY'],
             'TEST_PARALLEL_DATA_NAME': os.environ['TEST_PARALLEL_DATA_NAME'],
             'TEST_PARALLEL_DATA': os.environ['TEST_PARALLEL_DATA'],
-            'REGION': os.environ['MIE_REGION'],
-            'MIE_STACK_NAME': os.environ['MIE_STACK_NAME'],
+            'REGION': os.environ['REGION'],
+            'MI_STACK_NAME': os.environ['MI_STACK_NAME'],
             'ACCESS_KEY': os.environ['AWS_ACCESS_KEY_ID'],
             'SECRET_KEY': os.environ['AWS_SECRET_ACCESS_KEY']
             }
@@ -46,7 +46,7 @@ def stack_resources(testing_env_variables):
     # are the workflow api and testing stubs present?
 
     client = boto3.client('cloudformation', region_name=testing_env_variables['REGION'])
-    response = client.describe_stacks(StackName=testing_env_variables['MIE_STACK_NAME'])
+    response = client.describe_stacks(StackName=testing_env_variables['MI_STACK_NAME'])
     outputs = response['Stacks'][0]['Outputs']
 
     for output in outputs: